remove migration code from SDK and add it as a Sherlock state proxy
